+++ example/pubspec.yaml
@@ -3,7 +3,7 @@ description: Demonstrates how to use the sheet_music plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
