+++ example/test/widget_test.dart
@@ -13,13 +13,13 @@ import 'package:sheet_music_example/main.dart';
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(MyApp());
+    await tester.pumpWidget(SheetMusicExample());
 
     // Verify that platform version is retrieved.
     expect(
       find.byWidgetPredicate(
         (Widget widget) =>
-            widget is Text && widget.data.startsWith('Running on:'),
+            widget is Text && widget.data!.startsWith('Running on:'),
       ),
       findsOneWidget,
     );
